+++ cmd/add.js
@@ -25,63 +25,23 @@ checkConfigFile();
 var projectConfig = require('../project_config');
 
 // project html file path
-var htmlFilePath = projectConfig.basePaths.buildRoot + '/dev/html/' + moduleName + '.html';
+var htmlFilePath = projectConfig.basePaths.buildRoot + '/src/' + moduleName + '/index.html';
 // project js file path
-var jsFilePath =  projectConfig.basePaths.buildRoot + '/src/js/' + moduleName + '.js';
-// project css file path
-var cssFilePath = projectConfig.basePaths.buildRoot + '/src/css/' + moduleName + '.css';
+var jsFilePath =  projectConfig.basePaths.buildRoot + '/src/' + moduleName + '/index.js';
 
 // source html file path
 var sourceHtmlFilePath = vars.lilaRoot + '/project_files/demo/base.html';
 // source js file path
 var sourceJsFilePath = vars.lilaRoot + '/project_files/demo/base.js';
-// source css file path
-var sourceCssFilePath = vars.lilaRoot + '/project_files/demo/base.css';
 
-/**
- * module directory level
- *
- * example:
- *     index -> 1
- *     test/index -> 2
- *     parent/test/index -> 3
- */
-var modulePathLevels = moduleName.split('/').length;
-/**
- * module relative prefix
- *
- * example:
- *     index -> ../
- *     test/index -> ../../
- *     parent/test/index -> ../../../
- *
- */
-var moduleRelativePrefix = _.repeat('../', modulePathLevels);
-
-var data = {
-    module: moduleName,
-    cssPath: moduleRelativePrefix + 'css/' + moduleName,
-    jsPath: moduleRelativePrefix + 'js/' + moduleName
-};
-
-if (fs.existsSync(htmlFilePath) || fs.existsSync(jsFilePath) || fs.existsSync(cssFilePath)) {
+if (fs.existsSync(htmlFilePath) || fs.existsSync(jsFilePath)) {
     logger.error('Module ' + moduleName + ' has already been added.');
     process.exit(0);
 }
 
-// make css file
-fsExtra.outputFileSync(cssFilePath, fsExtra.readFileSync(sourceCssFilePath, 'utf8'));
 // make js file
-var jsTmpContent = fsExtra.readFileSync(sourceJsFilePath, 'utf8');
-_.forEach(data, (value, key) => {
-    jsTmpContent = jsTmpContent.replace(new RegExp('{{' + key + '}}', 'g'), value);
-});
-fsExtra.outputFileSync(jsFilePath, jsTmpContent);
+fsExtra.outputFileSync(jsFilePath, fsExtra.readFileSync(sourceJsFilePath, 'utf8'));
 // make html file
-var htmlTmpContent = fsExtra.readFileSync(sourceHtmlFilePath, 'utf8');
-_.forEach(data, (value, key) => {
-    htmlTmpContent = htmlTmpContent.replace(new RegExp('{{' + key + '}}', 'g'), value);
-});
-fsExtra.outputFileSync(htmlFilePath, htmlTmpContent);
+fsExtra.outputFileSync(htmlFilePath, fsExtra.readFileSync(sourceHtmlFilePath, 'utf8'));
 
 logger.success('Lila add module ' + moduleName + ' successful!');